requirements.txt
Django==1.5.1
Markdown==2.3.1
MySQL-python==1.2.4
PIL==1.1.7
Pillow==2.0.0
South==0.7.6
argparse==1.2.1
django-avatar==1.0.5
django-bootstrap-toolkit==2.11.5
django-classy-tags==0.4
django-common-helpers==0.4
django-countries==1.5
django-cron==0.2.8
django-dajax==0.9.2
django-dajaxice==0.5.5
django-extensions==1.1.1
django-filer==0.9.4
django-google-charts==0.1.1
django-haystack==2.0.0-beta
django-mockups==0.4.8
django-mptt==0.5.5
django-notification==0.2
django-pagination==1.0.7
django-polymorphic==0.4.2
django-profiles==0.2
django-qsstats-magic==0.7.2
django-registration==0.8
django-sekizai==0.6.1
django-threaded-messages==0.1.36
django-timedeltafield==0.7.0
django-uni-form==0.9.0
easy-thumbnails==1.2
haystack==0.15
pygraphviz==1.1
python-dateutil==2.1
python-ptrace==0.6.4
simplejson==3.1.3
six==1.3.0
sorl-thumbnail==11.12
wiki==0.0.19
wsgiref==0.1.2
